@eddyw 2022-11-06
Closing in favor of #265 

After fixing a couple of issues I haven't considered, the performance isn't _that_ impressive anymore (at least to justify the +80 bytes) lol - PR #265 performs better and it's more stable across different use case than the changes in this PR e.g: the simple case benchmark performs slightly worst here and most benefits seem to be seen _only_ with big arrays while PR #265 performance gains can be seen in simple cases as well as computed/effect with thousand sources.